+++ python/triton/compiler/__init__.py
@@ -1,2 +1,2 @@
-from .code_generator import CompilationError
 from .compiler import CompiledKernel, compile
+from .errors import CompilationError
