+++ python/triton/compiler/code_generator.py
@@ -8,6 +8,7 @@ import triton
 import triton._C.libtriton.triton as _triton
 from .. import impl
 from .. import language as tl
+from . import errors
 
 
 def mangle_ty(ty):
@@ -907,55 +908,5 @@ class CodeGenerator(ast.NodeVisitor):
                 except Exception as e:
                     message = "<failed to evaluate assertion message: " + repr(e) + ">"
 
-            raise CompileTimeAssertionFailure(None, node, message)
+            raise errors.CompileTimeAssertionFailure(None, node, message)
         return None
-
-
-class CompilationError(Exception):
-    source_line_count_max_in_message = 12
-
-    def _format_message(self) -> str:
-        node = self.node
-        if self.src is None:
-            source_excerpt = " <source unavailable>"
-        else:
-            source_excerpt = self.src.split('\n')[:node.lineno][-self.source_line_count_max_in_message:]
-            if source_excerpt:
-                source_excerpt.append(' ' * node.col_offset + '^')
-                source_excerpt = '\n'.join(source_excerpt)
-            else:
-                source_excerpt = " <source empty>"
-
-        message = "at {}:{}:{}".format(node.lineno, node.col_offset, source_excerpt)
-        if self.error_message:
-            message += '\n' + self.error_message
-        return message
-
-    def __init__(self, src: Optional[str], node: ast.AST, error_message: Union[str, 'triton.language.core.constexpr', None]):
-        self.src = src
-        self.node = node
-        self.error_message = _unwrap_if_constexpr(error_message)
-        self.message = self._format_message()
-
-    def set_source_code(self, src: Optional[str]):
-        self.src = src
-        self.message = self._format_message()
-
-    def __str__(self):
-        return self.message
-
-    def __repr__(self):
-        return "{}({!r})".format(type(self).__name__, self.message)
-
-    def __reduce__(self):
-        # this is necessary to make CompilationError picklable
-        return type(self), (self.src, self.node, self.error_message)
-
-
-class CompileTimeAssertionFailure(CompilationError):
-    """Specific exception for failed tests in `static_assert` invocations"""
-    pass
-
-
-class UnsupportedLanguageConstruct(CompilationError):
-    pass
